Highlight window: support minimized windows
also
- replace QHash::operator[]
- replace double by float (-> ARM, and we don't need that precision)
- weaken some value checks ("== 1.0f")
- fix last paint of hiding windows (w->addRepaintFull() trap)

REVIEW: 103404

+++ effects/highlightwindow/highlightwindow.h
@@ -50,8 +50,8 @@ private:
 
     bool m_finishing;
 
-    double m_fadeDuration;
-    QHash<EffectWindow*, double> m_windowOpacity;
+    float m_fadeDuration;
+    QHash<EffectWindow*, float> m_windowOpacity;
 
     long m_atom;
     QList<EffectWindow*> m_highlightedWindows;
